+++ src/argilla/client/feedback/training/schemas.py
@@ -407,7 +407,7 @@ class TrainingTask:
             [Dict[str, Any]], Union[None, Tuple[str, str, str, str], Iterator[Tuple[str, str, str, str]]]
         ],
     ) -> "TrainingTaskForChatCompletion":
-        """Training data for chat comletion
+        """Training data for chat completion
         Args:
             formatting_func: A formatting function converting a dictionary of records into zero,
                 one or more chat-turn-role-content text tuples.
@@ -1346,7 +1346,7 @@ class TrainingTaskForChatCompletionFormat(BaseModel):
 
 
 class TrainingTaskForChatCompletion(BaseModel, TrainingData):
-    """Training data for chat comletion
+    """Training data for chat completion
 
     Args:
         formatting_func: A formatting function converting a dictionary of records into zero,